+++ src/cli/actions/clear.rs
@@ -16,7 +16,7 @@ impl Runner {
 
         // check command and flags
         if self.args.params().len() > 2 {
-            return self.invalid_cmd_err();
+            return self.invalid_args_err(1);
         }
         self.check_flags(
             "run",