+++ src/cli/actions/tree.rs
@@ -25,7 +25,7 @@ impl Runner {
 
         // checks
         if self.args.params().len() > 2 {
-            return self.invalid_cmd_err();
+            return self.invalid_args_err(1);
         }
         self.check_flags(
             "tree",